decouple hash utils
