+++ src/utils/index.ts
@@ -1,2 +1,8 @@
 export { extractMetaMaskSigner } from "./metamask.js";
-export { concatenate } from "./bytes.js";
+export {
+  concatenate,
+  writeUIntLE,
+  buildBigIntFromUint8Array,
+  zeroPadLE,
+} from "./bytes.js";
+export { sha256, poseidonHash } from "./hash.js";
